Load default action steps from configuration (#253)
* Move steps to jbi.actions.steps

* Rename test_default to test_steps

* Move test_default_with_assignee_and_status into test_steps

* Load default action steps from configuration

* Add tests for the default action behaviour

* Rename groups of steps. create -> new, update -> existing

* Remove default value config/config.prod.yaml


* Move remapping of steps into helper

* Remove useless pylint annotation

* Adjust docstrings

* Merge unspecified groups with default ones

* Update secrets baseline
