+++ .secrets.baseline
@@ -122,9 +122,9 @@
         "filename": "README.md",
         "hashed_secret": "04e78d6e804f2b59e6cb282cb9ed2c7bfd8a9737",
         "is_verified": false,
-        "line_number": 218
+        "line_number": 242
       }
     ]
   },
-  "generated_at": "2022-08-30T10:02:55Z"
+  "generated_at": "2022-09-28T11:04:17Z"
 }
